+++ src/shared/data/element-selectors.js
@@ -9,11 +9,24 @@ export const elementSelectors = {
     old: "#ytpdc-playlist-summary-old",
     new: "#ytpdc-playlist-summary-new",
   },
-  playlistMetadata: {
-    old: "ytd-playlist-sidebar-renderer #items",
-    new: ".immersive-header-content .metadata-action-bar",
-    youtubePremium: ".yt-flexible-actions-view-model-wiz__action-row",
-  },
+  playlistMetadata: [
+    {
+      selector: ".yt-page-header-view-model__page-header-content",
+      queryMethod: "querySelectorAllAndFilter",
+    },
+    {
+      selector: ".immersive-header-content .metadata-action-bar",
+      queryMethod: "querySelector",
+    },
+    {
+      selector: "ytd-playlist-sidebar-renderer #items",
+      queryMethod: "querySelector",
+    },
+    {
+      selector: ".yt-flexible-actions-view-model-wiz__action-row",
+      queryMethod: "querySelector",
+    },
+  ],
   video: "ytd-playlist-video-renderer",
   playlist: "ytd-playlist-video-list-renderer #contents",
   channelName: ".ytd-channel-name",